swarm/state: refactor InmemoryStore
